Fix linking to not look for a system libstdc++.
Not sure what changed (I think an upstream LLVM change, but maybe
a Linux distro change), but several folks have been running into
problems finding a standard C++ library when running the Carbon link
step. It was actually breaking our example build when LLD didn't find
the right libstdc++ install, but it finds one reliably on our build bots
and sometimes for some of the developers.

For now, just remove the C++ standard library from the link. We're not
doing that level of interop, and the plan is really to do that not with
the system standard C++ library but by building and bundling libc++ with
the installed toolchain.

Left a comment explaining what's going on here.