Update the conference talk list for EuroLLVM videos (#4114)
Also reorder links so that the future/most recent items are first,
putting older videos lower down.

Noting the lightning talk because it's there, but also that it's a
lightning talk because most of the others are 40m/1h, so 5m is very
short by comparison (I was tempted to label the length of everything,
but not sure how valuable others view that).